- Updated to Python 3.8 so renamed __self->self etc. - filter_custom_vjp now checks that all non-vjp_args are not differentiated.
